pyproject.toml
[project]
name = "afml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "joblib>=1.5.3",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "optuna>=4.7.0",
    "pandas>=2.3.3",
    "polars>=1.0.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "ruff>=0.14.13",
]
